Add non-broken implementation of pySerial read()
This commit copies the implementation of seria.Serial.read() and
removes one line from it: "if timeout.expired():". With this change,
the function provides classical POSIX semantics of read(), which
returns as soon as some bytes are received. The implementation in
pySerial either waits for the full buffer or timeout expiration -
nothing in between. This means that with pySerial, you either pay big
overherd by reading character-by-character or have increased latency
due to waiting.

Unfortunately, there is no way to fix this in pySerial without
changing the semantics and possibly breaking some applications or
introducing a fixed function with a different name. Also, it seems
that pySerial is no loger maintained (last commit 16 months ago, PRs
without maintainer responses).

It would be much simpler not to use pySerial and use /dev/ttyXXX
directly. This project does not need the features offered by pySerial
(mainly abort read from another thread). The only problem of not using
pySerial would be lower portability.